requirements.txt
Flask
Werkzeug
pdfplumber
python-docx
easyocr
spacy
sentence-transformers
torch
Pillow